pyproject.toml
[project]
name = "dns-bench"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dnspython>=2.7.0",
    "rich>=14.0.0",
]
